+++ app/server/appsmith-interfaces/src/main/java/com/appsmith/external/helpers/SmartSubstitutionHelper.java
@@ -1,0 +1,22 @@
+package com.appsmith.external.helpers;
+
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
+
+public class SmartSubstitutionHelper {
+
+    public static String replaceQuestionMarkWithDollarIndex(String query) {
+        final AtomicInteger counter = new AtomicInteger();
+        String updatedQuery = query.chars()
+                .mapToObj(c -> {
+                    if (c == '?') {
+                        return "$" + counter.incrementAndGet();
+                    }
+
+                    return Character.toString(c);
+                })
+                .collect(Collectors.joining());
+
+        return updatedQuery;
+    }
+}
